LayoutTests/platform/glib/imported/w3c/web-platform-tests/trusted-types/require-trusted-types-for-expected.txt

CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Refused to connect to http://web-platform.test:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Refused to connect to http://web-platform.test:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: This requires a TrustedScriptURL value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Refused to connect to http://web-platform.test:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.

PASS Require trusted types for 'script' block create HTML.
PASS Require trusted types for 'script' block create script.
PASS Require trusted types for 'script' block create script URL.
PASS Set require trusted types for 'script' without CSP for trusted types don't block policy creation and using.
